requirements.txt
# JARVIS V4 — Full Requirements
# Install all: pip install -r requirements.txt

# ── Core ──────────────────────────────────────────────────────────────────────
requests>=2.31.0
python-dotenv>=1.0.0

# ── Terminal UI ───────────────────────────────────────────────────────────────
rich>=13.0.0

# ── System Monitoring ─────────────────────────────────────────────────────────
psutil>=5.9.0
GPUtil>=1.4.0

# ── Voice ─────────────────────────────────────────────────────────────────────
SpeechRecognition>=3.10.0
pyttsx3>=2.90
# pyaudio (microphone backend):
#   Windows: pip install pipwin && pipwin install pyaudio
#   Linux:   sudo apt install python3-pyaudio
#   Mac:     brew install portaudio && pip install pyaudio

# ── Whisper (Offline Speech-to-Text) ──────────────────────────────────────────
# Uncomment to enable offline voice recognition
# openai-whisper>=20231117
# torch>=2.2.0

# ── REST API ──────────────────────────────────────────────────────────────────
flask>=3.0.0

# ── Clipboard ────────────────────────────────────────────────────────────────
pyperclip>=1.8.2

# ── GUI Automation ───────────────────────────────────────────────────────────
pyautogui>=0.9.54
pygetwindow>=0.0.9

# ── Global Hotkeys ───────────────────────────────────────────────────────────
keyboard>=0.13.5

# ── Encryption / Security ────────────────────────────────────────────────────
cryptography>=41.0.0

# ── TOTP (2FA Codes) ─────────────────────────────────────────────────────────
pyotp>=2.9.0

# ── QR Code Generation ───────────────────────────────────────────────────────
qrcode[pil]>=7.4.2

# ── Image Processing / OCR ───────────────────────────────────────────────────
Pillow>=10.0.0
pytesseract>=0.3.10
# Install Tesseract OCR binary separately:
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
# Linux:   sudo apt install tesseract-ocr

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python>=4.8.0

# ── PDF Tools ────────────────────────────────────────────────────────────────
PyPDF2>=3.0.0

# ── Web Scraping ─────────────────────────────────────────────────────────────
beautifulsoup4>=4.12.0

# ── YouTube Download ─────────────────────────────────────────────────────────
yt-dlp>=2023.12.30

# ── Music Recognition ────────────────────────────────────────────────────────
shazamio>=0.4.0

# ── Internet Speed Test ──────────────────────────────────────────────────────
speedtest-cli>=2.1.3

# ── WHOIS Lookup ─────────────────────────────────────────────────────────────
python-whois>=0.8.0

# ── Smart Home / IoT ─────────────────────────────────────────────────────────
paho-mqtt>=1.6.1

# ── Telegram Bot (Optional) ──────────────────────────────────────────────────
# python-telegram-bot>=20.0

# ── YAML Support ─────────────────────────────────────────────────────────────
pyyaml>=6.0

# ── AI Integration (OpenRouter-Compatible) ───────────────────────────────────
# Included for future extensibility
openai>=1.0.0

# ── Minimal Install (Core Features Only) ─────────────────────────────────────
# pip install requests rich psutil flask pyperclip pyautogui cryptography python-dotenv